### 2019-10-16 #1803 @majensen
I've reached out to a couple of colleagues via email who may not be familiar with JOSS -- but I will @ -ping them here as well: @seandavi and @dcgenomics.
