### 2019-10-26 #1837 @whedon
[REVIEW]: CoralP: Flexible visualization of the human phosphatome
**Submitting author:** @dphansti (<a href="http://orcid.org/0000-0003-2123-0051">Douglas Phanstiel</a>)
**Repository:** <a href="https://github.com/PhanstielLab/coralp" target ="_blank">https://github.com/PhanstielLab/coralp</a>
**Version:** v1.0.1
