### 2019-10-30 #1837 @stulacy
Hi!

Could you please clean up some of the typos in the paper, such as: 'inlcuding', the odd symbol in the first sentence of Methods, and capitalising 'JavaScript'.
Also, the first two sentences in Methods are somewhat duplicated.
