### 2019-10-30 58da2e8
fixed several typos
