+++ paper/paper.md
@@ -45,11 +45,11 @@ Contact: douglas_phanstiel@med.unc.edu
 
 Protein phosphatases are members of a large protein family that regulate a wide array of biological processes via the removal of phosphoryl groups from their substrates [@Chen2017-bb; @Tonks2006-sq]. Like their counterparts, kinases (which add phosphoryl groups to substrates), phosphatases have become attractive targets for pharmaceutical development [@Cohen2002-yj]. As such, there exists a great need for methods to analyze, interpret, and communicate experimental results within the context of the entire human phosphatome.
 
-Numerous methods have been developed to visualize the human kinome [@Metz2018-ad; @Eid2017-gf; @Chartier2013-es], inlcuding our kinase visualization tool, Coral, an interactive web application that produces publication-quality visualizations of the human kinome. Coral allows for the simultaneous encoding of three attributes per kinase, supports three modes of visualization, and produces high-resolution vector images. Recently, Chen et al. published a phylogenetic phosphatase tree [@Chen2017-bb]; however, to the best of our knowledge, there are currently no available tools to visualize phosphatase data within the context of this or any other format. Here, we describe CoralP, the first and to the best of our knowledge only web application for the generation of customizable, publication-quality visualizations of the human phosphatome.
+Numerous methods have been developed to visualize the human kinome [@Metz2018-ad; @Eid2017-gf; @Chartier2013-es], including our kinase visualization tool, Coral, an interactive web application that produces publication-quality visualizations of the human kinome. Coral allows for the simultaneous encoding of three attributes per kinase, supports three modes of visualization, and produces high-resolution vector images. Recently, Chen et al. published a phylogenetic phosphatase tree [@Chen2017-bb]; however, to the best of our knowledge, there are currently no available tools to visualize phosphatase data within the context of this or any other format. Here, we describe CoralP, the first and to the best of our knowledge only web application for the generation of customizable, publication-quality visualizations of the human phosphatome.
 
 # Methods
 
-CoralP is adapted fr¬om Coral and is written using R and Javascript. CoralP is written in R and javascript and makes use of the following libraries: shiny [@Chang2018-jb], shinydashboard [@Chang2018-ef], shinyBS [Bailey2015-ft], shinyWidgets [@Perrier2019-io], readr [@Wickham2018-zb], rsvg [@Ooms2018-tr], RColorBrewer [@Neuwirth2014-ux], and D3.js [@Bostock2011-je]. The phosphatome tree was adapted from Chen et al. and manually redrawn using vector graphics in Adobe Illustrator [@Chen2017-bb]. 
+CoralP was adapted from Coral and was written using R and JavaScript miking use of the following libraries: shiny [@Chang2018-jb], shinydashboard [@Chang2018-ef], shinyBS [Bailey2015-ft], shinyWidgets [@Perrier2019-io], readr [@Wickham2018-zb], rsvg [@Ooms2018-tr], RColorBrewer [@Neuwirth2014-ux], and D3.js [@Bostock2011-je]. The phosphatome tree was adapted from Chen et al. and manually redrawn using vector graphics in Adobe Illustrator [@Chen2017-bb]. 
 
 # Results
 
