### 2019-10-31 57fb1f0
fixed typos
