+++ paper/paper.md
@@ -53,7 +53,7 @@ CoralP was adapted from Coral and was written using R and JavaScript making use 
 
 # Results
 
-CoralP is built using the same underlying framework as Coral and therefore makes use of similar strategies for data input, setting selection, and data download. Phosphatase attributes can be entered using pull-down menus or pasted into text boxes. CoralP supports multiple identifiers including UniProt, ENSEMBL, Entrez, and HGNC. Visualization options including color palettes, data scales, fonts, and protein identifiers can all be edited via an intuitive graphical user interface. CoralP was written using the reactive programming package Shiny [@Chang2018-jb], so CoralP plots update in real-time as users adjust settings, making it easy to generate a plot perfectly customized to the user's preferences.
+CoralP is built using the same underlying framework as Coral and therefore makes use of similar strategies for data input, setting selection, and data download. Phosphatase attributes can be entered using pull-down menus or pasted into text boxes. CoralP supports multiple identifiers including UniProt, ENSEMBL, Entrez, and HGNC. Visualization options including color palettes, data scales, fonts, and protein identifiers can all be edited via an intuitive graphical user interface. CoralP was written using the reactive programming package Shiny [@Chang2018-jb], so CoralP plots update in real-time as users adjust settings, making it easy to generate plots perfectly customized to the user's preferences.
 
 CoralP offers highly flexible methods for visualizing qualitative and quantitative phosphatase attributes within the context of the entire protein family (**Figure 1**). Data can be encoded via multiple modalities and displayed via either tree or network views. Qualitative and quantitative phosphatase attributes can be encoded through three separate modalities: branch color, node color, and node size. The resulting data can be displayed either as a phosphatase tree (modified from Chen et al.) or as a network organized in a radial or force-directed fashion.
 
