### 2019-10-31 dfe182d
removed repetitive use of the word 'modalities'
