### 2019-10-31 #1837 @dphansti
@stulacy Thanks for finding these issues. We have corrected those typos and made several other tiny edits to improve clarity and accuracy.
