### 2019-11-05 PhanstielLab/coralp#1 @daissi
Installation instructions
Part of JOSS review (https://github.com/openjournals/joss-reviews/issues/1837):

Could you add installation instructions? There is a list of R packages dependencies but no information about the installation of CoralP itself. Would be good as well to convert it to an R package [1-2].

[1] https://cran.r-project.org/doc/manuals/r-release/R-exts.html
[2] http://r-pkgs.had.co.nz/
