### 2019-11-05 c694e9f
expand section on running locally
