### 2019-11-05 PhanstielLab/coralp#1 @dphansti
Thanks so much for your feedback. We agree that it would be helpful to provide more detailed instructions for running the code locally. So we have added a new section to the README called “Running Locally” that provides information and code for downloading the code, installing prerequisites, and running the app locally.

I am not sure if it makes sense to turn it into an R package. Packages are typically collections of functions. But Coralp is a Shiny application that has been designed to run through a browser. As such Shiny apps are not typically distributed as packages. To do so would be unusual and wouldn’t be any easier than simply following the instructions to run locally.
