### 2019-11-12 PhanstielLab/coralp#2 @daissi
Typo in paper
Part of JOSS review (openjournals/joss-reviews#1837):

Could you fix the legend of figure 1? i.e.: "Figure 1: Figure1. ....", there is a duplication of "Figure 1".

There is also a problem with the reference Bailey2015-ft which is not considered during the generation of the pdf.
