### 2019-11-12 171b45d
Fix reference and figure caption.
