### 2019-11-13 #1837 @daissi
> @daissi, @stulacy - how are the checks coming? Can I ask you to check off or create issues for @dphansti for these over the next week? Let me know a timeline that works for you.

Everything is fine from my side! :thumbsup:
