### 2019-11-13 #1837 @stulacy
The only thing left I need to do is try and run the package locally which I
will try and complete tonight.

The only thing I'm unsure about is how to assess the "Automated Tests"
point, since this is a web-app rather than a package interfaced with via an
API. Is this criteria not valid for web-apps or should they still have some
sort of testing procedure?
