### 2019-11-19 #1837 @stulacy
I've gone through the remaining points and am happy that the submission meets them all, although I still am unsure how to handle the Automated Tests criteria as mentioned above. Any thoughts @majensen ?
