### 2019-11-22 #1837 @majensen
@dphansti  - per @stulacy comments: I agree that there should be some tests associated with the R code (the JS is less critical IMO). I wish I had time to help you stub these out, but for now let me suggest you look at https://testthat.r-lib.org/ - this looks really easy to use, and is a common testing system in other R-based pubs in JOSS. Here's an example : https://github.com/junbinzhao/FluxCalR/tree/master/tests
I can provide more advice if needed - but I wanted to get you looking at this.
thanks!
