### 2019-11-27 e9ad2a4
Adding Travis CI config file.
