+++ .travis.yml
@@ -1,0 +1,6 @@
+language: R
+sudo: false
+cache: packages
+
+script:
+ - R -r R/tests.R
