### 2019-11-27 ed483f4
Adding DESCRIPTION for Travis CI.
