+++ .travis.yml
@@ -2,7 +2,5 @@ language: R
 sudo: false
 cache: packages
 
-install:
-
 script:
  - R -r R/tests.R
