### 2019-11-27 70f63c6
Adding shinytest.
