+++ DESCRIPTION
@@ -6,20 +6,21 @@ Author: Amit Min, Erika Deoudes, Marielle L. Bond, Eric Davis, Doug Phanstiel
 Maintainer: Doug Phanstiel <douglas_phanstiel@med.unc.edu>
 Description: Flexible visualization of the human phosphatome
 Imports:
-    shiny, 
-	shinydashboard, 
-	shinyBS, 
-	readr, 
-	rsvg, 
-	shinyWidgets, 
-	RColorBrewer, 
-	svgPanZoom, 
-	colourpicker, 
-	DT, 
-	data.tree, 
-	jsonlite
+  shiny, 
+  shinydashboard, 
+  shinyBS, 
+  readr, 
+  rsvg, 
+  shinyWidgets, 
+  RColorBrewer, 
+  svgPanZoom, 
+  colourpicker, 
+  DT, 
+  data.tree, 
+  jsonlite
 Suggests:
-    testthat, 
-	here, 
-	grDevices
+  testthat, 
+  shinytest, 
+  here, 
+  grDevices
 License: MIT + file LICENSE
