+++ .travis.yml
@@ -1,6 +1,7 @@
 language: R
 sudo: false
 cache: packages
+os: osx
 
 script:
  - R -f run_tests.R
