+++ .travis.yml
@@ -2,6 +2,11 @@ language: R
 sudo: false
 cache: packages
 os: osx
+addons:
+  homebrew:
+    packages:
+    - phantomjs
+    update: true
 
 script:
  - R -f run_tests.R
