### 2019-11-27 fdfb5e3
Revert "Trying with compare images."
This reverts commit 56b53c1c4429b721446ca4120fef47149fbdeb67.
