### 2019-11-27 89a6c74
Adding Travis CI badge.
