### 2019-12-02 #1837 @dphansti
@daissi, @stulacy, @majensen We have added unit tests using the `testthat` package for key functions, as well as an application-level test using the `shinytest` package that emulates simple tasks in the user interface and compares with pre-validated results. We also configured Travis CI to automatically run these tests with every change and report failures.
