### 2019-12-02 #1837 @stulacy
Great, thanks @dphansti . I've completed my review.
