### 2019-12-03 #1837 @majensen
@dphansti thanks very much for that additional work. Thanks @stulacy and @daissi for the review!
Next step is to create a public archive of the software with a DOI. @dphansti, you can use [Zenodo](zenodo.org) for this purpose, or other service you prefer. If you're unfamiliar with this process, have a look at https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729 for a brief tutorial. Once this is done, please report the DOI back here.
thanks again
