+++ README.md
@@ -1,5 +1,7 @@
 [![Travis build status](https://travis-ci.org/PhanstielLab/coralp.svg?branch=master)](https://travis-ci.org/PhanstielLab/coralp)
 
+[![DOI](https://zenodo.org/badge/179145702.svg)](https://zenodo.org/badge/latestdoi/179145702)
+
 # Overview
 
 http://phanstiel-lab.med.unc.edu/coralp/
