### 2019-12-03 c80a43c
Explicitly set color palettes to sequential by default.
