### 2019-12-03 401c7ce
Revert "Explicitly set color palettes to sequential by default."
This reverts commit c80a43cd105337f41bb56d418b7595cd0e1a4d78.
