### 2019-12-03 #1837 @majensen
@openjournals/joss-eics this paper is recommended for publication. Please initiate final checks- thanks!
