### 2019-12-03 #1837 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
